Bound calculation improvements

+++ simulant/window_base.cpp
@@ -298,7 +298,7 @@ void WindowBase::run_update() {
     if(frame_counter_time_ >= 1.0) {
         stats->set_frames_per_second(frame_counter_frames_);
 
-        frame_time_in_milliseconds_ = 1000.0 / double(frame_counter_frames_);
+        frame_time_in_milliseconds_ = 1000.0 / float(frame_counter_frames_);
         frame_counter_frames_ = 0;
         frame_counter_time_ = 0.0;
     }
